requirements.txt
absl-py==0.11.0
asgiref==3.3.1
astunparse==1.6.3
cachetools==4.2.0
chardet==4.0.0
click==7.1.2
cython==0.29.21
decorator==4.4.2
django==3.1.5
flask==1.1.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
image==1.5.33
imageio==2.9.0
importlib-metadata==3.4.0
imutils==0.5.4
ipython-genutils==0.2.0
ipyvolume==0.5.2
ipywebrtc==0.6.0
itsdangerous==1.1.0
jinja2==2.11.3
joblib==1.0.0
keras-preprocessing==1.1.2
keras-tuner==1.0.0
markdown==3.3.3
markupsafe==1.1.1
matplotlib==3.3.4
mrcnn==0.2
networkx==2.5
nibabel==2.5.2
nii2png==0.2.7
nilearn==0.5.2
niwidgets==0.2.2
numpy==1.16.6
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.8
pandas==1.2.4
pandocfilters==1.4.3
pillow==8.1.0
plotly==4.14.3
protobuf==3.14.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
pytz==2020.5
pywavelets==1.1.1
pywin32==300
pywinpty==1.1.1
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7
scikit-image==0.18.1
scikit-learn==0.20.4
scipy==1.6.0
six==1.15.0
sqlparse==0.4.1
tabulate==0.8.7
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
terminado==0.10.0
terminaltables==3.1.0
threadpoolctl==2.1.0
tifffile==2021.1.14
tqdm==4.56.0
urllib3==1.26.2
webencodings==0.5.1
werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0